Add files via upload
This is the zipped file of the data, you need to unzip/extract it before using it for analysis. * Reminder: This data is a little bit different from original ones, and are only for teaching purposes. If you want to use the WVS data for research, please refer to official website and apply for their data: https://www.worldvaluessurvey.org/WVSContents.jsp.